fix: some tests work now! gotta clean and iterate further
